Move megacontrol views to admin
