+++ control/views.py
@@ -1,10 +1,6 @@
 from django import forms
-from django.contrib import messages
-from django.contrib.admin.views.decorators import staff_member_required
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.http import JsonResponse
-from django.utils.decorators import method_decorator
-from django.utils.safestring import mark_safe
 from django.views import View
 from django.views.generic import DetailView, CreateView, RedirectView, TemplateView
 from django.views.generic.detail import SingleObjectMixin
@@ -12,7 +8,6 @@ from django.views.generic.detail import SingleObjectMixin
 from actstream import action
 from sendfile import sendfile
 
-from control.questionnaire_duplicate import QuestionnaireDuplicateMixin
 from .docx import generate_questionnaire_file
 from .models import Questionnaire, QuestionFile, ResponseFile, Control
 
@@ -154,48 +149,3 @@ class QuestionnaireDetail(LoginRequiredMixin, WithListOfControlsMixin, DetailVie
         if not self.request.user.profile.is_inspector:
             queryset = queryset.filter(is_draft=False)
         return queryset
-
-
-@method_decorator(staff_member_required, name='dispatch')
-class MegacontrolConfirm(QuestionnaireDuplicateMixin, DetailView):
-    template_name = "ecc/megacontrol_confirm.html"
-    context_object_name = 'questionnaire'
-
-    def get_queryset(self):
-        return Questionnaire.objects.all()
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['controls_to_copy_to'] = self.get_controls_to_copy_to(self.object)
-        return context
-
-
-@method_decorator(staff_member_required, name='dispatch')
-class Megacontrol(LoginRequiredMixin, QuestionnaireDuplicateMixin, SingleObjectMixin, RedirectView):
-    url = '/admin/control/questionnaire/'
-    model = Questionnaire
-
-    def get_queryset(self):
-        return Questionnaire.objects.all()
-
-    def get(self, *args, **kwargs):
-        questionnaire = self.get_object()
-        controls_to_copy_to = self.get_controls_to_copy_to(questionnaire)
-
-        created_questionnaires = []
-        for control_to_copy_to in controls_to_copy_to:
-            created_questionnaire = self.copy_questionnaire(questionnaire, control_to_copy_to)
-            created_questionnaires.append(created_questionnaire)
-
-        message = 'Vous avez créé les questionnaires suivants : <ul>'
-        for created_questionnaire in created_questionnaires:
-            message += f'<li>'
-            message += f'  <a href="/admin/control/questionnaire/{created_questionnaire.id}/change/">'
-            message += f'    <b> {created_questionnaire.id} : {created_questionnaire} </b>'
-            message += f'  </a>'
-            message += f'  dans l\'espace <b>{ created_questionnaire.control }</b>'
-            message += f'</li>'
-        message += '</ul>'
-        messages.success(self.request, mark_safe(message))
-
-        return super().get(*args, **kwargs)
